Create EmailAddressBook
This reads lines from a tab delimited text file, splits each line element into an array, adds the split elements as properties into a Person object. Adds the person object to a list, adds the name of each person to a listbox and then creates a second form to display all the information from the selected person.