Create MultiFormLogin
Project that creates an account and saves it to tab delimited txt file. Login form that reads the information from the txt file to match it with the entered information to grant access. If access is granted the list of users w/ all their information is shown in a listbox on another form.